account details: first part is over

+++ src/layouts/ReadyAccount.jsx
@@ -1,21 +1,28 @@
-import React, { useEffect } from 'react'
+import React, { useEffect, useState } from 'react'
 import { useNavigate, useParams } from 'react-router-dom'
 import HandleFormSubmit from '../helpers/handleForSubmit'
 
 function ReadyAccount() {
+  const [confirm, setConfirm] = useState(true)
   const navigate = useNavigate()
   const { token } = useParams()
 
-  const { error, isPending, handleSubmit } = HandleFormSubmit({})
+  const { error, handleSubmit } = HandleFormSubmit({})
 
   useEffect(() => {
     handleSubmit({}, 'user/create', { token })
       .then((resp) => {
+        if (!resp) {
+          setConfirm(true)
+          return
+        }
         console.log({ resp })
         localStorage.setItem('token', resp.data.token)
+        setConfirm(false)
       })
       .catch(() => {
         console.error('Error en el inicio de sesión:')
+
         return
       })
   }, [])
@@ -25,7 +32,7 @@ function ReadyAccount() {
       ReadyAccount
       {error.error && <p className='error'>{error.msg}</p>}
       <button
-        disabled={isPending}
+        disabled={confirm}
         onClick={() => navigate('/create-account/payment ')}
       >
         Continue -&gt;
